Build when run, so config is available
If we want to modify UI config we need to do so using environment variables. These are written into the static files at build time. So if we want to allow these variables to set when we create the container then we need the build to happen when we create the container. That's what this change is for. See the docker yml files in stroom-resources for where these are now set.